# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_document_route_api_v1_matters__matter_id__documents_post": {
        "properties": {
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "expected_revision",
          "file"
        ],
        "title": "Body_upload_document_route_api_v1_matters__matter_id__documents_post",
        "type": "object"
      },
...
      "CreateMatterRequest": {
        "properties": {
          "eligibility_confirmed": {
            "title": "Eligibility Confirmed",
            "type": "boolean"
          },
          "eligibility_version": {
            "default": "self_single_v1.0.0",
            "title": "Eligibility Version",
            "type": "string"
          }
        },
        "required": [
          "eligibility_confirmed"
        ],
        "title": "CreateMatterRequest",
        "type": "object"
      },
      "GenerationResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "final_confirmed": {
            "title": "Final Confirmed",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "matter_id": {
            "title": "Matter Id",
            "type": "string"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "matter_id",
          "revision",
          "status",
          "final_confirmed",
          "preview_url",
          "download_url",
          "sha256",
          "created_at"
        ],
        "title": "GenerationResponse",
        "type": "object"
      },
      "GenerationStartResponse": {
        "properties": {
          "generation": {
            "$ref": "#/components/schemas/GenerationResponse"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "generation",
          "job_id"
        ],
        "title": "GenerationStartResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobResponse": {
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "attempt",
          "max_attempts",
          "progress",
          "error_code",
          "result"
        ],
        "title": "JobResponse",
        "type": "object"
      },
...
      "MatterResponse": {
        "properties": {
          "confirmations": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Confirmations",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_title": {
            "title": "Display Title",
            "type": "string"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/MatterDocumentResponse"
            },
            "title": "Documents",
            "type": "array"
          },
          "eligibility_confirmed": {
            "title": "Eligibility Confirmed",
            "type": "boolean"
          },
          "eligibility_version": {
            "title": "Eligibility Version",
            "type": "string"
          },
          "facts": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Facts",
            "type": "object"
          },
          "generated_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Revision"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latest_generation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Generation Id"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "sources": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "title": "Sources",
            "type": "object"
          },
          "title_state": {
            "title": "Title State",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "validated_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated Revision"
          },
          "workflow_profile": {
            "title": "Workflow Profile",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workflow_profile",
          "eligibility_version",
          "eligibility_confirmed",
          "revision",
          "title_state",
          "display_title",
          "facts",
          "confirmations",
          "sources",
          "documents",
          "validated_revision",
          "generated_revision",
          "latest_generation_id",
          "created_at",
          "updated_at"
        ],
        "title": "MatterResponse",
        "type": "object"
      },
      "RevisionRequest": {
        "properties": {
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "RevisionRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Case Filing Assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/generations/{generation_id}": {
      "get": {
        "operationId": "get_generation_api_v1_generations__generation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "generation_id",
            "required": true,
            "schema": {
              "title": "Generation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Generation"
      }
    },
    "/api/v1/generations/{generation_id}/confirm": {
      "post": {
        "operationId": "confirm_generation_api_v1_generations__generation_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "generation_id",
            "required": true,
            "schema": {
              "title": "Generation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Generation"
      }
    },
    "/api/v1/generations/{generation_id}/download": {
      "get": {
        "operationId": "download_generation_api_v1_generations__generation_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "generation_id",
            "required": true,
            "schema": {
              "title": "Generation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Generation"
      }
    },
    "/api/v1/generations/{generation_id}/preview": {
      "get": {
        "operationId": "preview_generation_api_v1_generations__generation_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "generation_id",
            "required": true,
            "schema": {
              "title": "Generation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Generation"
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/api/v1/matters": {
      "get": {
        "operationId": "list_matters_api_v1_matters_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatterResponse"
                  },
                  "title": "Response List Matters Api V1 Matters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Matters"
      },
      "post": {
        "operationId": "create_matter_route_api_v1_matters_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Matter Route"
      }
    },
    "/api/v1/matters/{matter_id}": {
      "get": {
        "operationId": "get_matter_route_api_v1_matters__matter_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Matter Route"
      }
    },
    "/api/v1/matters/{matter_id}/documents": {
      "post": {
        "operationId": "upload_document_route_api_v1_matters__matter_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_route_api_v1_matters__matter_id__documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Document Route"
      }
    },
    "/api/v1/matters/{matter_id}/facts": {
      "put": {
        "operationId": "save_facts_route_api_v1_matters__matter_id__facts_put",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFactsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Facts Route"
      }
    },
    "/api/v1/matters/{matter_id}/generations": {
      "post": {
        "operationId": "start_generation_api_v1_matters__matter_id__generations_post",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Generation"
      }
    },
    "/api/v1/matters/{matter_id}/validate": {
      "post": {
        "operationId": "validate_route_api_v1_matters__matter_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Route"
      }
    },
...
  }
}